requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.1.1
django-bootstrap5==24.2
djangorestframework==3.15.2
et-xmlfile==1.1.0
fastapi==0.114.2
gunicorn==23.0.0
h11==0.14.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.38.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
